+++ packages/test-utils/index.js
@@ -1,10 +1,10 @@
 const pFinally = require('p-finally');
 const supertest = require('supertest-light');
 const extractStack = require('extract-stack');
-const { Keystone } = require('@voussoir/keystone');
-const { WebServer } = require('@voussoir/server');
-const { MongooseAdapter } = require('@voussoir/adapter-mongoose');
-const { KnexAdapter } = require('@voussoir/adapter-knex');
+const { Keystone } = require('@keystone-alpha/keystone');
+const { WebServer } = require('@keystone-alpha/server');
+const { MongooseAdapter } = require('@keystone-alpha/adapter-mongoose');
+const { KnexAdapter } = require('@keystone-alpha/adapter-knex');
 const MongoDBMemoryServer = require('mongodb-memory-server').default;
 
 function setupServer({ name, adapterName, createLists = () => {} }) {