+++ packages/fields/types/Color/index.js
@@ -5,7 +5,7 @@ module.exports = {
   type: 'Color',
   implementation: Text,
   views: {
-    Controller: path.join(__dirname, '../Text/Controller'),
+    Controller: path.join(__dirname, '../Text/views/Controller'),
     Field: path.join(__dirname, './views/Field'),
     Cell: path.join(__dirname, './views/Cell'),
     Filter: path.join(__dirname, '../Text/views/Filter'),