+++ test-projects/twitter-login/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @keystone-alpha/cypress-project-twitter-login
 
+## 1.0.3
+
+- [patch][de616f7e](https://github.com/keystonejs/keystone-5/commit/de616f7e):
+
+  - Use new authStrategy APIs
+
+- Updated dependencies [9a9f214a](https://github.com/keystonejs/keystone-5/commit/9a9f214a):
+- Updated dependencies [de616f7e](https://github.com/keystonejs/keystone-5/commit/de616f7e):
+- Updated dependencies [4ed35dfd](https://github.com/keystonejs/keystone-5/commit/4ed35dfd):
+  - @keystone-alpha/keystone@1.0.3
+  - @keystone-alpha/admin-ui@3.0.0
+  - @keystone-alpha/fields@3.0.0
+  - @keystone-alpha/core@2.0.0
+  - @keystone-alpha/server@2.0.0
+
 ## 1.0.2
 
 - [patch][7417ea3a](https://github.com/keystonejs/keystone-5/commit/7417ea3a):